Add Prettier
I've noticed that some of our files are beginning to get difficult to
work with due to formatting issues. With several people collaborating on
this project now, we've got a mixture of tabs and spaces, different
indent sizes, etc. Let's use Prettier (with a default config for now) to
keep things looking tidy.