Remove non-Astro Code
I've been working on a re-write of ConfessIt in
[Astro](https://astro.build/). I chose to start this as a new Astro
project to get a clean start in the new framework. I've been tracking
that work in a separate git repository. Although the rewrite is not yet
complete, I think I've made enough progress that I'm confident in the
direction I want to head and I'm also ready to merge that work back into
this repository. Doing so will be a little tricky since these
repositories don't currently share history. There are a few different
ways I could do this but here's what I've chosen.

1. [THIS COMMIT] Delete all files that have been re-written or won't be
used in the astro branch. (But keep files that will be reused like
translations.)
2. Rebase the astro repo onto this (now mostly empty) repo, in a branch
named 'astro'.
3. (Soon), merge the astro branch into main in this repo, to officially
use astro in production.

For files that won't change (like translations), this perserves history.
For many React components, we'll (at least partially) lose some history
(as the files will be removed and then reappear in a slightly different
location), but I think that's fine and it allows me to preserve the
history of the Astro re-write.