Refresh the About ConfessIt Page
Back in 2023, I realized that because ConfessIt was a client-side
rendered React SPA at the time, it had very poor SEO. I created a simple
HTML landing page as a quick solution:
https://github.com/kas-catholic/www.confessit.app

Since then, we've re-written ConfessIt using Astro. One of the benefits
of this is that Astro can do static site generation, so we get much
better SEO with HTML served by our server, except for the few islands of
interactivity. Another benefit is that all of this is translated, the
same way we translate the app itself.

So, my plan is to get rid of the old English-only landing page, and
redirect www.confessit.app to confessit.app/about. Before doing that, I
want to make confessit.app/about look a little nicer, and ensure all the
important content from the old landing page is present on the new site.
To accomplish this, I'm also adding a short landing page for the
ConfessIt Android app, and splitting off "About Confession" to its own
page.